==> analysis/R_analysis/scr/00_libraries.R <==
renv::install("tidyverse")
renv::install("bioc::BiocManager")
renv::install("bioc::MOFA2")
renv::install("R.utils")
renv::install("GGally")
renv::install("ggpubr")
renv::install("bioc::MOFAdata")
renv::install("psych")
renv::install("randomForest")
renv::install("survminer")
renv::install("patchwork")
renv::install("bioc::msigdbr")
renv::install("bioc::multiGSEA")
renv::install("bioc::org.Hs.eg.db")
renv::install("bioc::metaboliteIDmapping")
